Explicitly set an epoch
Per packaging experts, we need to explicitly set an epoch flag for rhosp
consumption of the correct (and older) version of the train packaging,
as opposed to the latest master branch version upstream.

This is related to https://review.rdoproject.org/r/c/rdoinfo/+/39303
which pinned the train version to the branch, as versions >3.0.0 of
ironic-python-agent-builder remove support for iscsi based deployments
of baremetal nodes. However, in Train, this is still supported.

As such, putting the epoch in place.

Resolves: rhbz#2057169
Related: rhbz#2050539
Change-Id: I1b24b59f8d6ef4752833882cd8b38f28c038a008